fix: allow fractional orbs in ActiveAspect and ChartAspectSetting
ActiveAspect.orb and _ChartAspectSetting.orb were typed as int, but the
rest of the codebase (AspectModel.orbit, axis_orb_limit, get_orb(), etc.)
already uses float. Pydantic v2 strictly enforces int fields, rejecting
valid fractional orb values like 7.5 with int_from_float errors.

Bump to 6.0.0a28.

+++ kerykeion/schemas/kr_models.py
@@ -895,7 +895,7 @@ class RelationshipScoreModel(SubscriptableBaseModel):
 
 class ActiveAspect(TypedDict):
     name: AspectName
-    orb: int
+    orb: float
 
 
 class TransitMomentModel(SubscriptableBaseModel):
